refactor(markerscope): improve error handling and code organization
